Bugfix: Detect correct Cargo project
Previously, `find_project_dir` returned "." whenever a Cargo.toml was
present—without checking whether that Cargo project matched the
requested name. So if we asked for 'some_project', the app treated
the current rustr directory as the target and executed any commands.
This breaks the requirement that we cannot run rustr on itself.

This commit fixes the bug by checking verifying that if we are in a
Rust project directory, the current project in Cargo.toml matches the
requested name. If we aren't in the directory of the project that is
specified, we use the canonical project location.

This ensures that we always use the actual project directory for a
specified project name, or the current directory if no project name
is specified. In all cases, we will not run commands on the rustr
application itself.